Add option to sync node power state from DB
Previously, during _sync_node_power_state, when a node's power state did
not match the expected state, the DB was always updated and a warning
logged.

This patch causes the node's power state to instead be set to the
expected value, by default, and adds a CONF option to enable the old
behavior.

Implements bp:keep-powered-off-nodes-off

Co-Authored-By: Michael Davies <michael@the-davies.net>
Change-Id: I1f5af0061dbe8c57b8bd3f8a65b215aadd549551