Add logical name support to Ironic
Everything that is tracked in Ironic is done via UUID. This isn't very
friendly to humans.

This change adds a new concept to ironic, that being the <logical name>,
which can be used interchangeably with the <node uuid>.  Everywhere a
<node uuid> can be specified, we are able to instead specify a <logical
name>, if such an association exists for that node.

APIImpact:
    This raises the X-OpenStack-Ironic-API-Version to 1.5, and
    adds a new "name" property to nodes, which can be substituted
    for the UUID in requests.

    This also adds a 'node' parameter for 'GET /v1/nodes/validate'. This
    parameter can be a node's UUID or name. The existing 'node_uuid'
    parameter is still supported.

DocImpact

blueprint: logical-names

Change-Id: I505181a1df064194848e3d5b79d01746024ce037