Disable XML now that we have WSME/Pecan support
We've had a desire to disable XML from the Ironic API for some
time, however to do so needed support in WSME/Pecan.  That's now
been added, so we can implement the Ironic side.

To do so, without modifying every single wsexpose decorator in
Ironic (i.e. every single endpoint) we monkey patch wsexpose,
adding in the specification of the 'rest_content_types' that we
wish to support and then invoke the original wsexpose decorator.

Change-Id: I94d0387722225c4356c867f63b6f1a61356f317d
Closes-bug: 1271317